refactor: separate to presentational and container component for Edit
